**Before**
Unregistering user from an event was done entirely in views. This means that a function(s) in views both cleaned form data and ran logic that checked if unregistration was possible, thereby interacting with `models.py`. 

**Now**
Form data cleaning is now done by `EventUnregisterForm` in `forms.py`, and unregistration logic is now implemented as a static method `remove_atendee_from_code` from `Event` class. It takes `user` and `code` (which is now validated by a form) as arguments, and checks if unregistration is possible. If it is not, it raises a `ValidationError`, which is catched by a function in `views.py` that displays this error's message. If unregistration is possible, it is performed and user is informed.

**Additionaly**

- cosmetic changes to code, name unification, improved readability
- removed files from magicevents/media/images by adding this folder to gitignore, but folder itself is kept
- added run and setup scripts
- created todos

+++ magicevents/rootapp/models.py
@@ -1,9 +1,14 @@
 from django.db import models
 from django.contrib.auth.models import User
 from django.core.exceptions import ValidationError
+from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
 
 from random import randint
 import datetime
+
+
+# QUESTION: where and how to put remove_atendee? it does not use self so...
+#           staticmethod? classmethod?
 
 class Event(models.Model):
     title = models.CharField(max_length=100)
@@ -27,22 +32,34 @@ class Event(models.Model):
                                          user=user,
                                          registration_code=code)
 
-    def remove_atendee(self, user, code):
-        registration = EventRegistration.objects.get(user=user,
-                                                    event=self,
-                                                    registration_code=code)
-        registration.delete()
+    @staticmethod
+    def remove_atendee_from_code(user, code):
+        try:
+            registration = EventRegistration.objects.get(user=user,
+                                                         registration_code=code)
+        except MultipleObjectsReturned:
+            raise ValidationError('Uncanny.')
+        except ObjectDoesNotExist:
+            raise ValidationError('This code does not match any registrations.')
 
+        event = registration.event
 
-    def get_days_to_start(self):
+        # check if this event starts after two days
         today = datetime.date.today()
-        diff_days = (self.start_date - today).days
-        return diff_days if diff_days >= 0 else -1
+        days_difference = (event.start_date - today).days
 
-    def get_duration(self):
-        return (self.end_date - self.start_date).days
+        if days_difference < 2:
+            raise ValidationError('You cannot cancel an event after two days \
+                                  before the start date')
 
+        # check if this event's duration does not exceed two days
+        event_duration = (event.end_date - event.start_date).days
 
+        if event_duration > 2:
+            raise ValidationError('You cannot cancel an event that is longer \
+                                  than two days.')
+
+        registration.delete()
 
 
 class EventRegistration(models.Model):
@@ -50,10 +67,20 @@ class EventRegistration(models.Model):
                               verbose_name='Event',
                               on_delete=models.CASCADE)
     user = models.ForeignKey(User,
-                            verbose_name='Attendee',
-                            on_delete=models.CASCADE)
+                             verbose_name='Attendee',
+                             on_delete=models.CASCADE)
 
     registration_code = models.PositiveIntegerField(
                                 default=randint(111111, 999999))
+
+    class Meta:
+        verbose_name = "Event registration"
+        verbose_name_plural = "Event registrations"
+
     def __str__(self):
         return f'User {self.user.username} registered for {self.event.title}'
+
+    @staticmethod
+    def already_registered(event, user):
+        return bool(EventRegistration.objects.filter(event=event,
+                                                     user=user).count())
